requirements.txt
pandas == 1.1.1
google-cloud-videointelligence == 1.16.0
opencv-python == 4.4.0.44
google-cloud-vision == 2.0.0
pillow == 7.2.0
tqdm == 4.51.0
pytube == 9.7.0